refactor: Clean up API constants, and explain usage
